pyproject.toml
[project]
name = "cortex"
version = "0.1.0"
description = "Semantic intelligence pipeline for natural language data access"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.3.0",
    "google-cloud-aiplatform>=1.75.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-discoveryengine>=0.13.0",
    "neo4j>=5.25.0",
    "faiss-cpu>=1.8.0",
    "lkml>=1.3.0",
    "pyyaml>=6.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "python-dotenv>=1.0.0",
    # SafeChain LLM access (Amex internal — install from internal PyPI)
    "safechain",
    "ee_config",
    # LangChain pinned — SafeChain requires this exact version
    "langchain-core==0.3.83",
    # MCP protocol
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
